Refactor some modules in the directory 'common' of FElib.

+++ FElib/src/common/scale_time_manager.F90
@@ -83,7 +83,7 @@ module scale_time_manager
   
   !-----------------------------------------------------------------------------
 
-  real(DP), private :: TIME_RES_RESUME
+  integer, private :: TIME_RES_RESUME
 
 contains
